requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
scipy
faiss-cpu
langchain_chroma
streamlit
langchain_ollama
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
